refactor: Issueコンポーネントを分割した
